Move location of file
