+++ app/pages/project/instances/instance/tabs/NetworkingTab.tsx
@@ -24,7 +24,6 @@ import { IpGlobal24Icon, Networking24Icon } from '@oxide/design-system/icons/rea
 import { AttachEphemeralIpModal } from '~/components/AttachEphemeralIpModal'
 import { AttachFloatingIpModal } from '~/components/AttachFloatingIpModal'
 import { HL } from '~/components/HL'
-import { IpLink } from '~/components/IpLink'
 import { CreateNetworkInterfaceForm } from '~/forms/network-interface-create'
 import { EditNetworkInterfaceForm } from '~/forms/network-interface-edit'
 import { getInstanceSelector, useInstanceSelector, useProjectSelector } from '~/hooks'
@@ -39,6 +38,7 @@ import { Table } from '~/table/Table'
 import { Badge } from '~/ui/lib/Badge'
 import { CreateButton } from '~/ui/lib/CreateButton'
 import { EmptyMessage } from '~/ui/lib/EmptyMessage'
+import { IpLink } from '~/ui/lib/IpLink'
 import { TableControls, TableEmptyBox, TableTitle } from '~/ui/lib/Table'
 import { TipIcon } from '~/ui/lib/TipIcon'
 import { pb } from '~/util/path-builder'
